requirements.txt
pytube==12.1.0
tqdm ==4.64.1
colorama==0.4.5
Pillow==9.2.0
customtkinter==5.1.2
requests==2.27.1
